requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
coverage==6.4.4
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
flake8==5.0.4
flake8-django==1.1.5
importlib-metadata==4.12.0
iniconfig==1.1.1
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
zipp==3.8.1
